Rm eddsa verifier from jose, add validation utils
