+++ identity_credential/Cargo.toml
@@ -31,6 +31,7 @@ thiserror.workspace = true
 url = { version = "2.4", default-features = false }
 
 [dev-dependencies]
+identity_eddsa_verifier = { version = "=0.7.0-alpha.7", path = "../identity_eddsa_verifier", default-features = false, features = ["ed25519"] }
 iota-crypto = { version = "0.23", default-features = false, features = ["ed25519", "std", "random"] }
 proptest = { version = "1.0.0", default-features = false, features = ["std"] }
 serde_json.workspace = true
@@ -47,6 +48,6 @@ default = ["revocation-bitmap", "validator", "credential", "presentation", "doma
 credential = []
 presentation = ["credential"]
 revocation-bitmap = ["dep:dataurl", "dep:flate2", "dep:roaring"]
-validator = ["dep:itertools", "dep:serde_repr", "credential", "presentation", "identity_verification/eddsa"]
+validator = ["dep:itertools", "dep:serde_repr", "credential", "presentation"]
 domain-linkage = ["validator"]
 domain-linkage-fetch = ["domain-linkage", "dep:reqwest", "dep:futures"]
