+++ identity_credential/src/validator/vc_jwt_validation/mod.rs
@@ -6,8 +6,10 @@ mod credential_jwt_validation_options;
 mod credential_jwt_validator;
 mod decoded_jwt_credential;
 mod error;
+mod jwt_credential_validator_utils;
 
 pub use credential_jwt_validation_options::*;
 pub use credential_jwt_validator::*;
 pub use decoded_jwt_credential::*;
 pub use error::*;
+pub use jwt_credential_validator_utils::*;
