+++ identity_credential/src/validator/vp_jwt_validation/mod.rs
@@ -5,8 +5,10 @@ mod decoded_jwt_presentation;
 mod error;
 mod jwt_presentation_validation_options;
 mod jwt_presentation_validator;
+mod jwt_presentation_validator_utils;
 
 pub use decoded_jwt_presentation::*;
 pub use error::*;
 pub use jwt_presentation_validation_options::*;
 pub use jwt_presentation_validator::*;
+pub use jwt_presentation_validator_utils::*;
