+++ identity_jose/Cargo.toml
@@ -20,11 +20,6 @@ subtle = { version = "2.5", default-features = false }
 thiserror.workspace = true
 zeroize = { version = "1.6", default-features = false, features = ["std", "zeroize_derive"] }
 
-[features]
-# Enables jws verification based on the EdDSA algorithm.
-eddsa = ["iota-crypto/ed25519"]
-default = ["eddsa"]
-
 [dev-dependencies]
 anyhow = "1"
 iota-crypto = { version = "0.23", features = ["ed25519", "random", "hmac"] }
